created bullet VFX
Added gun muzzle flash as well as a hit effect vfx that instantiates a bullet impact particle effect wherever the player has shot. this uses raycasts to pinpoint the location of the effect instantiation. also created a self destruct script and attached it to instantiated particle effect gameobjects and moved all instantiated bullet effects to a "spawnatruntime" gameobject.